+++ Backend.sln
@@ -20,6 +20,14 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "src\Domain\Domain
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntegrationTests", "Tests\IntegrationTests\IntegrationTests.csproj", "{1799BF13-598F-4799-A0D0-36FED87C17BD}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "core", "core", "{7A47FB0B-2211-4007-B1E3-4EA04A31A748}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "infrastructure", "infrastructure", "{BEB8A8A7-AF73-407F-8D36-6B61565F4C04}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tools", "tools", "{8E0839A6-7A18-4F9E-890D-3DE4C8DFF0BD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tools", "src\Tools\Tools.csproj", "{7A8670AA-500F-46DB-84D2-1ADE0B018603}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -54,14 +62,22 @@ Global
 		{1799BF13-598F-4799-A0D0-36FED87C17BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1799BF13-598F-4799-A0D0-36FED87C17BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1799BF13-598F-4799-A0D0-36FED87C17BD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7A8670AA-500F-46DB-84D2-1ADE0B018603}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7A8670AA-500F-46DB-84D2-1ADE0B018603}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7A8670AA-500F-46DB-84D2-1ADE0B018603}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7A8670AA-500F-46DB-84D2-1ADE0B018603}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{61AE7D38-24E7-4E5E-8B24-84C5181D9936} = {D5F16979-2055-4ADA-87E5-119A4759663D}
-		{DD35E752-6F46-4845-87CF-17D6DBD0E36B} = {A868740B-AEE8-4EDA-8BAE-31E988E0FBE1}
 		{388A165A-1758-486F-BA81-F4AC7DC24F6C} = {B4B407C0-D7B3-4F5A-A8D0-8808EC0D6CBA}
-		{F6F65830-8898-42FE-A881-0E86BD516A47} = {A868740B-AEE8-4EDA-8BAE-31E988E0FBE1}
-		{21147C06-4506-40C7-BF07-46C324FF4CF9} = {A868740B-AEE8-4EDA-8BAE-31E988E0FBE1}
-		{F608F5C5-85F1-47EB-B3EB-B3DCB7B8AA7E} = {A868740B-AEE8-4EDA-8BAE-31E988E0FBE1}
 		{1799BF13-598F-4799-A0D0-36FED87C17BD} = {B4B407C0-D7B3-4F5A-A8D0-8808EC0D6CBA}
+		{7A47FB0B-2211-4007-B1E3-4EA04A31A748} = {A868740B-AEE8-4EDA-8BAE-31E988E0FBE1}
+		{F608F5C5-85F1-47EB-B3EB-B3DCB7B8AA7E} = {7A47FB0B-2211-4007-B1E3-4EA04A31A748}
+		{21147C06-4506-40C7-BF07-46C324FF4CF9} = {7A47FB0B-2211-4007-B1E3-4EA04A31A748}
+		{BEB8A8A7-AF73-407F-8D36-6B61565F4C04} = {A868740B-AEE8-4EDA-8BAE-31E988E0FBE1}
+		{F6F65830-8898-42FE-A881-0E86BD516A47} = {BEB8A8A7-AF73-407F-8D36-6B61565F4C04}
+		{DD35E752-6F46-4845-87CF-17D6DBD0E36B} = {BEB8A8A7-AF73-407F-8D36-6B61565F4C04}
+		{8E0839A6-7A18-4F9E-890D-3DE4C8DFF0BD} = {7A47FB0B-2211-4007-B1E3-4EA04A31A748}
+		{7A8670AA-500F-46DB-84D2-1ADE0B018603} = {8E0839A6-7A18-4F9E-890D-3DE4C8DFF0BD}
 	EndGlobalSection
 EndGlobal
