+++ Tests/IntegrationTests/Authentication/RegisterAdminTests.cs
@@ -1,6 +1,5 @@
 ﻿using System.Net;
 using Domain.Entity;
-using Domain.utils;
 using Dto;
 using IntegrationTests.FakeDbSetup;
 using IntegrationTests.TestFactory;
