+++ Tests/IntegrationTests/Item/UpdateItemTests.cs
@@ -1,6 +1,5 @@
 ﻿using System.Net;
 using System.Text;
-using Domain.utils;
 using Dto;
 using IntegrationTests.FakeDbSetup;
 using IntegrationTests.TestFactory;
