+++ Tests/UnitTests/Domain/Item/CreateItemTests.cs
@@ -1,7 +1,6 @@
 ﻿using Domain.Entity;
 using Domain.item;
 using Domain.Repository;
-using Domain.utils;
 using Dto;
 using Moq;
 using UnitTests.Factory;
