+++ src/Domain/Entity/AdminEntity.cs
@@ -1,6 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.Text.RegularExpressions;
-using Domain.utils;
 
 namespace Domain.Entity;
