+++ src/Domain/Entity/BillingPartyEntity.cs
@@ -1,7 +1,6 @@
 ﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Text.RegularExpressions;
-using Domain.utils;
 
 namespace Domain.Entity;
