+++ src/Domain/Entity/ItemEntity.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using Domain.utils;
 
 namespace Domain.Entity;
