+++ src/Domain/GlobalUsings.cs
@@ -1,0 +1,1 @@
+﻿global using Tools;
