+++ src/Domain/auth/AuthenticationDomain.cs
@@ -1,8 +1,6 @@
-using System.Text.RegularExpressions;
 using Domain.converters;
 using Domain.Entity;
 using Domain.Repository;
-using Domain.utils;
 using Dto;
 
 namespace Domain.auth;
