+++ src/Domain/item/ItemDomain.cs
@@ -2,7 +2,6 @@
 using System.Runtime.InteropServices;
 using Domain.Entity;
 using Domain.Repository;
-using Domain.utils;
 using Dto;
 
 namespace Domain.item;
